Refactor server management into separate app
