+++ src/server_manager/web_app/app.ts
@@ -14,7 +14,6 @@
 
 import * as sentry from '@sentry/electron';
 import {EventEmitter} from 'eventemitter3';
-import * as semver from 'semver';
 
 import * as digitalocean_api from '../cloud/digitalocean_api';
 import * as errors from '../infrastructure/errors';
@@ -28,81 +27,12 @@ import {DisplayServer, DisplayServerRepository, makeDisplayServer} from './displ
 import {parseManualServerConfig} from './management_urls';
 
 import {AppRoot} from './ui_components/app-root.js';
-import {DisplayAccessKey, DisplayDataAmount, ServerView} from './ui_components/outline-server-view.js';
+import {DisplayAccessKey, ServerView} from './ui_components/outline-server-view.js';
+import {ServerManagementApp} from "./server_management_app";
 
 // The Outline DigitalOcean team's referral code:
 //   https://www.digitalocean.com/help/referral-program/
 const UNUSED_DIGITALOCEAN_REFERRAL_CODE = '5ddb4219b716';
-
-const CHANGE_KEYS_PORT_VERSION = '1.0.0';
-const DATA_LIMITS_VERSION = '1.1.0';
-const CHANGE_HOSTNAME_VERSION = '1.2.0';
-// Date by which the data limits feature experiment will be permanently added or removed.
-export const DATA_LIMITS_AVAILABILITY_DATE = new Date('2020-06-02');
-const MAX_ACCESS_KEY_DATA_LIMIT_BYTES = 50 * (10 ** 9);  // 50GB
-
-function dataLimitToDisplayDataAmount(limit: server.DataLimit): DisplayDataAmount|null {
-  if (!limit) {
-    return null;
-  }
-  const bytes = limit.bytes;
-  if (bytes >= 10 ** 9) {
-    return {value: Math.floor(bytes / (10 ** 9)), unit: 'GB'};
-  }
-  return {value: Math.floor(bytes / (10 ** 6)), unit: 'MB'};
-}
-
-function displayDataAmountToDataLimit(dataAmount: DisplayDataAmount): server.DataLimit|null {
-  if (!dataAmount) {
-    return null;
-  }
-  if (dataAmount.unit === 'GB') {
-    return {bytes: dataAmount.value * (10 ** 9)};
-  } else if (dataAmount.unit === 'MB') {
-    return {bytes: dataAmount.value * (10 ** 6)};
-  }
-  return {bytes: dataAmount.value};
-}
-
-// Compute the suggested data limit based on the server's transfer capacity and number of access
-// keys.
-async function computeDefaultAccessKeyDataLimit(
-    server: server.Server, accessKeys?: server.AccessKey[]): Promise<server.DataLimit> {
-  try {
-    // Assume non-managed servers have a data transfer capacity of 1TB.
-    let serverTransferCapacity: server.DataAmount = {terabytes: 1};
-    if (isManagedServer(server)) {
-      serverTransferCapacity = server.getHost().getMonthlyOutboundTransferLimit();
-    }
-    if (!accessKeys) {
-      accessKeys = await server.listAccessKeys();
-    }
-    let dataLimitBytes = serverTransferCapacity.terabytes * (10 ** 12) / (accessKeys.length || 1);
-    if (dataLimitBytes > MAX_ACCESS_KEY_DATA_LIMIT_BYTES) {
-      dataLimitBytes = MAX_ACCESS_KEY_DATA_LIMIT_BYTES;
-    }
-    return {bytes: dataLimitBytes};
-  } catch (e) {
-    console.error(`Failed to compute default access key data limit: ${e}`);
-    return {bytes: MAX_ACCESS_KEY_DATA_LIMIT_BYTES};
-  }
-}
-
-async function showHelpBubblesOnce(serverView: ServerView) {
-  if (!window.localStorage.getItem('addAccessKeyHelpBubble-dismissed')) {
-    await serverView.showAddAccessKeyHelpBubble();
-    window.localStorage.setItem('addAccessKeyHelpBubble-dismissed', 'true');
-  }
-  if (!window.localStorage.getItem('getConnectedHelpBubble-dismissed')) {
-    await serverView.showGetConnectedHelpBubble();
-    window.localStorage.setItem('getConnectedHelpBubble-dismissed', 'true');
-  }
-  if (!window.localStorage.getItem('dataLimitsHelpBubble-dismissed') &&
-      serverView.supportsAccessKeyDataLimit) {
-    await serverView.showDataLimitsHelpBubble();
-    window.localStorage.setItem('dataLimitsHelpBubble-dismissed', 'true');
-  }
-}
 
 function isManagedServer(testServer: server.Server): testServer is server.ManagedServer {
   return !!(testServer as server.ManagedServer).getHost;
@@ -110,10 +40,6 @@ function isManagedServer(testServer: server.Server): testServer is server.Manage
 
 function isManualServer(testServer: server.Server): testServer is server.ManualServer {
   return !!(testServer as server.ManualServer).forget;
-}
-
-function localizeDate(date: Date, language: string): string {
-  return date.toLocaleString(language, {year: 'numeric', month: 'long', day: 'numeric'});
 }
 
 type DigitalOceanSessionFactory = (accessToken: string) => digitalocean_api.DigitalOceanSession;
@@ -126,7 +52,8 @@ export class App {
   private serverBeingCreated: server.ManagedServer;
 
   constructor(
-      private appRoot: AppRoot, private readonly version: string,
+      private appRoot: AppRoot, private serverManagementApp: ServerManagementApp,
+      private readonly version: string,
       private createDigitalOceanSession: DigitalOceanSessionFactory,
       private createDigitalOceanServerRepository: DigitalOceanServerRepositoryFactory,
       private manualServerRepository: server.ManualServerRepository,
@@ -134,51 +61,24 @@ export class App {
       private digitalOceanTokenManager: TokenManager, private surveys: Surveys) {
     appRoot.setAttribute('outline-version', this.version);
 
+    // Managed server (i.e. DigitalOcean) related events
     appRoot.addEventListener('ConnectToDigitalOcean', (event: CustomEvent) => {
       this.connectToDigitalOcean();
     });
     appRoot.addEventListener('SignOutRequested', (event: CustomEvent) => {
       this.clearCredentialsAndShowIntro();
     });
-
     appRoot.addEventListener('SetUpServerRequested', (event: CustomEvent) => {
       this.createDigitalOceanServer(event.detail.regionId);
     });
-
     appRoot.addEventListener('DeleteServerRequested', (event: CustomEvent) => {
       this.deleteSelectedServer();
     });
-
     appRoot.addEventListener('ForgetServerRequested', (event: CustomEvent) => {
       this.forgetSelectedServer();
     });
-
-    appRoot.addEventListener('AddAccessKeyRequested', (event: CustomEvent) => {
-      this.addAccessKey();
-    });
-
-    appRoot.addEventListener('RemoveAccessKeyRequested', (event: CustomEvent) => {
-      this.removeAccessKey(event.detail.accessKeyId);
-    });
-
-    appRoot.addEventListener('RenameAccessKeyRequested', (event: CustomEvent) => {
-      this.renameAccessKey(event.detail.accessKeyId, event.detail.newName, event.detail.entry);
-    });
-
-    appRoot.addEventListener('SetAccessKeyDataLimitRequested', (event: CustomEvent) => {
-      this.setAccessKeyDataLimit(displayDataAmountToDataLimit(event.detail.limit));
-    });
-
-    appRoot.addEventListener('RemoveAccessKeyDataLimitRequested', (event: CustomEvent) => {
-      this.removeAccessKeyDataLimit();
-    });
-
-    appRoot.addEventListener('ChangePortForNewAccessKeysRequested', (event: CustomEvent) => {
-      this.setPortForNewAccessKeys(event.detail.validatedInput, event.detail.ui);
-    });
-
-    appRoot.addEventListener('ChangeHostnameForAccessKeysRequested', (event: CustomEvent) => {
-      this.setHostnameForAccessKeys(event.detail.validatedInput, event.detail.ui);
+    appRoot.addEventListener('CancelServerCreationRequested', (event: CustomEvent) => {
+      this.cancelServerCreation(this.selectedServer);
     });
 
     // The UI wants us to validate a server management URL.
@@ -224,14 +124,6 @@ export class App {
           });
     });
 
-    appRoot.addEventListener('EnableMetricsRequested', (event: CustomEvent) => {
-      this.setMetricsEnabled(true);
-    });
-
-    appRoot.addEventListener('DisableMetricsRequested', (event: CustomEvent) => {
-      this.setMetricsEnabled(false);
-    });
-
     appRoot.addEventListener('SubmitFeedback', (event: CustomEvent) => {
       const detail = event.detail;
       try {
@@ -245,14 +137,6 @@ export class App {
         console.error(`Failed to submit feedback: ${e}`);
         appRoot.showError(appRoot.localize('error-feedback'));
       }
-    });
-
-    appRoot.addEventListener('ServerRenameRequested', (event: CustomEvent) => {
-      this.renameServer(event.detail.newName);
-    });
-
-    appRoot.addEventListener('CancelServerCreationRequested', (event: CustomEvent) => {
-      this.cancelServerCreation(this.selectedServer);
     });
 
     appRoot.addEventListener('OpenImageRequested', (event: CustomEvent) => {
@@ -273,6 +157,64 @@ export class App {
     });
 
     onUpdateDownloaded(this.displayAppUpdateNotification.bind(this));
+
+    /**
+     * ServerManagementApp event listeners
+     */
+    // TODO: Move these event listeners to server_management_app once they no longer
+    //       depend on `selectedServer`.
+    // Server management events
+    appRoot.addEventListener('ServerRenameRequested', (event: CustomEvent) => {
+      serverManagementApp.renameServer(this.selectedServer, event.detail.newName).then(() => {
+        this.syncAndShowServer(this.selectedServer);
+      });
+    });
+    appRoot.addEventListener('ChangePortForNewAccessKeysRequested', (event: CustomEvent) => {
+      serverManagementApp.setPortForNewAccessKeys(
+        this.selectedServer, event.detail.validatedInput, event.detail.ui);
+    });
+    appRoot.addEventListener('ChangeHostnameForAccessKeysRequested', (event: CustomEvent) => {
+      serverManagementApp.setHostnameForAccessKeys(
+        this.selectedServer, event.detail.validatedInput, event.detail.ui);
+    });
+
+    // Access key events
+    appRoot.addEventListener('AddAccessKeyRequested', (event: CustomEvent) => {
+      serverManagementApp.addAccessKey(this.selectedServer);
+    });
+    appRoot.addEventListener('RemoveAccessKeyRequested', (event: CustomEvent) => {
+      serverManagementApp.removeAccessKey(this.selectedServer, event.detail.accessKeyId);
+    });
+    appRoot.addEventListener('RenameAccessKeyRequested', (event: CustomEvent) => {
+      serverManagementApp.renameAccessKey(
+        this.selectedServer, event.detail.accessKeyId, event.detail.newName, event.detail.entry);
+    });
+
+    // Metric events
+    appRoot.addEventListener('EnableMetricsRequested', (event: CustomEvent) => {
+      serverManagementApp.setMetricsEnabled(this.selectedServer, true);
+    });
+    appRoot.addEventListener('DisableMetricsRequested', (event: CustomEvent) => {
+      serverManagementApp.setMetricsEnabled(this.selectedServer, false);
+    });
+
+    // Data limits feature events
+    appRoot.addEventListener('SetAccessKeyDataLimitRequested', (event: CustomEvent) => {
+      serverManagementApp
+        .setAccessKeyDataLimit(
+          this.selectedServer,
+          ServerManagementApp.displayDataAmountToDataLimit(event.detail.limit))
+        .then((result) => {
+          if (result) {
+            this.surveys.presentDataLimitsEnabledSurvey();
+          }
+        });
+    });
+    appRoot.addEventListener('RemoveAccessKeyDataLimitRequested', (event: CustomEvent) => {
+      serverManagementApp.removeAccessKeyDataLimit(this.selectedServer).then(() => {
+        this.surveys.presentDataLimitsDisabledSurvey();
+      });
+    });
   }
 
   async start(): Promise<void> {
@@ -550,7 +492,8 @@ export class App {
       if (isHealthy) {
         // Sync the server display in case it was previously unreachable.
         this.syncServerToDisplay(server).then(() => {
-          this.showServer(server, displayServer);
+          this.displayServerRepository.storeLastDisplayedServerId(displayServer.id);
+          this.serverManagementApp.showServer(server, displayServer);
         });
       } else {
         // Display the unreachable server state within the server view.
@@ -815,155 +758,8 @@ export class App {
   private async syncAndShowServer(server: server.Server, timeoutMs = 250) {
     const displayServer = await this.syncServerToDisplay(server);
     await this.syncDisplayServersToUi();
-    this.showServer(server, displayServer);
-  }
-
-  // Show the server management screen. Assumes the server is healthy.
-  private async showServer(selectedServer: server.Server, selectedDisplayServer: DisplayServer) {
-    this.selectedServer = selectedServer;
-    this.appRoot.selectedServer = selectedDisplayServer;
-    this.displayServerRepository.storeLastDisplayedServerId(selectedDisplayServer.id);
-
-    // Show view and initialize fields from selectedServer.
-    const view = this.appRoot.getServerView(selectedDisplayServer.id);
-    view.isServerReachable = true;
-    view.serverId = selectedServer.getServerId();
-    view.serverName = selectedServer.getName();
-    view.serverHostname = selectedServer.getHostnameForAccessKeys();
-    view.serverManagementApiUrl = selectedServer.getManagementApiUrl();
-    view.serverPortForNewAccessKeys = selectedServer.getPortForNewAccessKeys();
-    view.serverCreationDate = localizeDate(selectedServer.getCreatedDate(), this.appRoot.language);
-    view.serverVersion = selectedServer.getVersion();
-    view.dataLimitsAvailabilityDate =
-        localizeDate(DATA_LIMITS_AVAILABILITY_DATE, this.appRoot.language);
-    view.accessKeyDataLimit = dataLimitToDisplayDataAmount(selectedServer.getAccessKeyDataLimit());
-    view.isAccessKeyDataLimitEnabled = !!view.accessKeyDataLimit;
-
-    const version = this.selectedServer.getVersion();
-    if (version) {
-      view.isAccessKeyPortEditable = semver.gte(version, CHANGE_KEYS_PORT_VERSION);
-      view.supportsAccessKeyDataLimit = semver.gte(version, DATA_LIMITS_VERSION);
-      view.isHostnameEditable = semver.gte(version, CHANGE_HOSTNAME_VERSION);
-    }
-
-    if (isManagedServer(selectedServer)) {
-      view.isServerManaged = true;
-      const host = selectedServer.getHost();
-      view.monthlyCost = host.getMonthlyCost().usd;
-      view.monthlyOutboundTransferBytes =
-          host.getMonthlyOutboundTransferLimit().terabytes * (10 ** 12);
-      view.serverLocation = this.getLocalizedCityName(host.getRegionId());
-    } else {
-      view.isServerManaged = false;
-    }
-
-    view.metricsEnabled = selectedServer.getMetricsEnabled();
-    this.appRoot.showServerView();
-    this.showMetricsOptInWhenNeeded(selectedServer, view);
-
-    // Load "My Connection" and other access keys.
-    try {
-      const serverAccessKeys = await selectedServer.listAccessKeys();
-      view.accessKeyRows = serverAccessKeys.map(this.convertToUiAccessKey.bind(this));
-      if (!view.accessKeyDataLimit) {
-        view.accessKeyDataLimit = dataLimitToDisplayDataAmount(
-            await computeDefaultAccessKeyDataLimit(selectedServer, serverAccessKeys));
-      }
-      // Show help bubbles once the page has rendered.
-      setTimeout(() => {
-        showHelpBubblesOnce(view);
-      }, 250);
-    } catch (error) {
-      console.error(`Failed to load access keys: ${error}`);
-      this.appRoot.showError(this.appRoot.localize('error-keys-get'));
-    }
-
-    this.showTransferStats(selectedServer, view);
-  }
-
-  private showMetricsOptInWhenNeeded(selectedServer: server.Server, serverView: ServerView) {
-    const showMetricsOptInOnce = () => {
-      // Sanity check to make sure the running server is still displayed, i.e.
-      // it hasn't been deleted.
-      if (this.selectedServer !== selectedServer) {
-        return;
-      }
-      // Show the metrics opt in prompt if the server has not already opted in,
-      // and if they haven't seen the prompt yet according to localStorage.
-      const storageKey = selectedServer.getServerId() + '-prompted-for-metrics';
-      if (!selectedServer.getMetricsEnabled() && !localStorage.getItem(storageKey)) {
-        this.appRoot.showMetricsDialogForNewServer();
-        localStorage.setItem(storageKey, 'true');
-      }
-    };
-
-    // Calculate milliseconds passed since server creation.
-    const createdDate = selectedServer.getCreatedDate();
-    const now = new Date();
-    const msSinceCreation = now.getTime() - createdDate.getTime();
-
-    // Show metrics opt-in once ONE_DAY_IN_MS has passed since server creation.
-    const ONE_DAY_IN_MS = 24 * 60 * 60 * 1000;
-    if (msSinceCreation >= ONE_DAY_IN_MS) {
-      showMetricsOptInOnce();
-    } else {
-      setTimeout(showMetricsOptInOnce, ONE_DAY_IN_MS - msSinceCreation);
-    }
-  }
-
-  private async refreshTransferStats(selectedServer: server.Server, serverView: ServerView) {
-    try {
-      const stats = await selectedServer.getDataUsage();
-      let totalBytes = 0;
-      // tslint:disable-next-line:forin
-      for (const accessKeyId in stats.bytesTransferredByUserId) {
-        totalBytes += stats.bytesTransferredByUserId[accessKeyId];
-      }
-      serverView.setServerTransferredData(totalBytes);
-
-      const accessKeyDataLimit = selectedServer.getAccessKeyDataLimit();
-      if (accessKeyDataLimit) {
-        // Make access key data usage relative to the data limit.
-        totalBytes = accessKeyDataLimit.bytes;
-      }
-
-      // Update all the displayed access keys, even if usage didn't change, in case the data limit
-      // did.
-      for (const accessKey of serverView.accessKeyRows) {
-        const accessKeyId = accessKey.id;
-        const transferredBytes = stats.bytesTransferredByUserId[accessKeyId] || 0;
-        let relativeTraffic =
-            totalBytes ? 100 * transferredBytes / totalBytes : (accessKeyDataLimit ? 100 : 0);
-        if (relativeTraffic > 100) {
-          // Can happen when a data limit is set on an access key that already exceeds it.
-          relativeTraffic = 100;
-        }
-        serverView.updateAccessKeyRow(accessKeyId, {transferredBytes, relativeTraffic});
-      }
-    } catch (e) {
-      // Since failures are invisible to users we generally want exceptions here to bubble
-      // up and trigger a Sentry report. The exception is network errors, about which we can't
-      // do much (note: ShadowboxServer generates a breadcrumb for failures regardless which
-      // will show up when someone explicitly submits feedback).
-      if (e instanceof errors.ServerApiError && e.isNetworkError()) {
-        return;
-      }
-      throw e;
-    }
-  }
-
-  private showTransferStats(selectedServer: server.Server, serverView: ServerView) {
-    this.refreshTransferStats(selectedServer, serverView);
-    // Get transfer stats once per minute for as long as server is selected.
-    const statsRefreshRateMs = 60 * 1000;
-    const intervalId = setInterval(() => {
-      if (this.selectedServer !== selectedServer) {
-        // Server is no longer running, stop interval
-        clearInterval(intervalId);
-        return;
-      }
-      this.refreshTransferStats(selectedServer, serverView);
-    }, statsRefreshRateMs);
+    this.displayServerRepository.storeLastDisplayedServerId(displayServer.id);
+    await this.serverManagementApp.showServer(server, displayServer);
   }
 
   private getS3InviteUrl(accessUrl: string, isAdmin = false) {
@@ -984,107 +780,6 @@ export class App {
       transferredBytes: 0,
       relativeTraffic: 0
     };
-  }
-
-  private addAccessKey() {
-    this.selectedServer.addAccessKey()
-        .then((serverAccessKey: server.AccessKey) => {
-          const uiAccessKey = this.convertToUiAccessKey(serverAccessKey);
-          this.appRoot.getServerView(this.appRoot.selectedServer.id).addAccessKey(uiAccessKey);
-          this.appRoot.showNotification(this.appRoot.localize('notification-key-added'));
-        })
-        .catch((error) => {
-          console.error(`Failed to add access key: ${error}`);
-          this.appRoot.showError(this.appRoot.localize('error-key-add'));
-        });
-  }
-
-  private renameAccessKey(accessKeyId: string, newName: string, entry: polymer.Base) {
-    this.selectedServer.renameAccessKey(accessKeyId, newName)
-        .then(() => {
-          entry.commitName();
-        })
-        .catch((error) => {
-          console.error(`Failed to rename access key: ${error}`);
-          this.appRoot.showError(this.appRoot.localize('error-key-rename'));
-          entry.revertName();
-        });
-  }
-
-  private async setAccessKeyDataLimit(limit: server.DataLimit) {
-    if (!limit) {
-      return;
-    }
-    const previousLimit = this.selectedServer.getAccessKeyDataLimit();
-    if (previousLimit && limit.bytes === previousLimit.bytes) {
-      return;
-    }
-    const serverView = this.appRoot.getServerView(this.appRoot.selectedServer.id);
-    try {
-      await this.selectedServer.setAccessKeyDataLimit(limit);
-      this.appRoot.showNotification(this.appRoot.localize('saved'));
-      serverView.accessKeyDataLimit = dataLimitToDisplayDataAmount(limit);
-      this.refreshTransferStats(this.selectedServer, serverView);
-      this.surveys.presentDataLimitsEnabledSurvey();
-    } catch (error) {
-      console.error(`Failed to set access key data limit: ${error}`);
-      this.appRoot.showError(this.appRoot.localize('error-set-data-limit'));
-      serverView.accessKeyDataLimit = dataLimitToDisplayDataAmount(
-          previousLimit || await computeDefaultAccessKeyDataLimit(this.selectedServer));
-      serverView.isAccessKeyDataLimitEnabled = !!previousLimit;
-    }
-  }
-
-  private async removeAccessKeyDataLimit() {
-    const serverView = this.appRoot.getServerView(this.appRoot.selectedServer.id);
-    try {
-      await this.selectedServer.removeAccessKeyDataLimit();
-      this.appRoot.showNotification(this.appRoot.localize('saved'));
-      this.refreshTransferStats(this.selectedServer, serverView);
-      this.surveys.presentDataLimitsDisabledSurvey();
-    } catch (error) {
-      console.error(`Failed to remove access key data limit: ${error}`);
-      this.appRoot.showError(this.appRoot.localize('error-remove-data-limit'));
-      serverView.isAccessKeyDataLimitEnabled = true;
-    }
-  }
-
-  private async setHostnameForAccessKeys(hostname: string, serverSettings: polymer.Base) {
-    this.appRoot.showNotification(this.appRoot.localize('saving'));
-    try {
-      await this.selectedServer.setHostnameForAccessKeys(hostname);
-      this.appRoot.showNotification(this.appRoot.localize('saved'));
-      serverSettings.enterSavedState();
-    } catch (error) {
-      this.appRoot.showError(this.appRoot.localize('error-not-saved'));
-      if (error.isNetworkError()) {
-        serverSettings.enterErrorState(this.appRoot.localize('error-network'));
-        return;
-      }
-      const message = error.response.status === 400 ? 'error-hostname-invalid' : 'error-unexpected';
-      serverSettings.enterErrorState(this.appRoot.localize(message));
-    }
-  }
-
-  private async setPortForNewAccessKeys(port: number, serverSettings: polymer.Base) {
-    this.appRoot.showNotification(this.appRoot.localize('saving'));
-    try {
-      await this.selectedServer.setPortForNewAccessKeys(port);
-      this.appRoot.showNotification(this.appRoot.localize('saved'));
-      serverSettings.enterSavedState();
-    } catch (error) {
-      this.appRoot.showError(this.appRoot.localize('error-not-saved'));
-      if (error.isNetworkError()) {
-        serverSettings.enterErrorState(this.appRoot.localize('error-network'));
-        return;
-      }
-      const code = error.response.status;
-      if (code === 409) {
-        serverSettings.enterErrorState(this.appRoot.localize('error-keys-port-in-use'));
-        return;
-      }
-      serverSettings.enterErrorState(this.appRoot.localize('error-unexpected'));
-    }
   }
 
   // Returns promise which fulfills when the server is created successfully,
@@ -1121,18 +816,6 @@ export class App {
         }
       });
     });
-  }
-
-  private removeAccessKey(accessKeyId: string) {
-    this.selectedServer.removeAccessKey(accessKeyId)
-        .then(() => {
-          this.appRoot.getServerView(this.appRoot.selectedServer.id).removeAccessKey(accessKeyId);
-          this.appRoot.showNotification(this.appRoot.localize('notification-key-removed'));
-        })
-        .catch((error) => {
-          console.error(`Failed to remove access key: ${error}`);
-          this.appRoot.showError(this.appRoot.localize('error-key-remove'));
-        });
   }
 
   private deleteSelectedServer() {
@@ -1188,34 +871,6 @@ export class App {
       this.showIntro();
       this.appRoot.showNotification(this.appRoot.localize('notification-server-removed'));
     });
-  }
-
-  private async setMetricsEnabled(metricsEnabled: boolean) {
-    try {
-      await this.selectedServer.setMetricsEnabled(metricsEnabled);
-      this.appRoot.showNotification(this.appRoot.localize('saved'));
-      // Change metricsEnabled property on polymer element to update display.
-      this.appRoot.getServerView(this.appRoot.selectedServer.id).metricsEnabled = metricsEnabled;
-    } catch (error) {
-      console.error(`Failed to set metrics enabled: ${error}`);
-      this.appRoot.showError(this.appRoot.localize('error-metrics'));
-    }
-  }
-
-  private async renameServer(newName: string) {
-    const view = this.appRoot.getServerView(this.appRoot.selectedServer.id);
-    try {
-      await this.selectedServer.setName(newName);
-      view.serverName = newName;
-      this.syncAndShowServer(this.selectedServer);
-    } catch (error) {
-      console.error(`Failed to rename server: ${error}`);
-      this.appRoot.showError(this.appRoot.localize('error-server-rename'));
-      const oldName = this.selectedServer.getName();
-      view.serverName = oldName;
-      // tslint:disable-next-line:no-any
-      (view.$.serverSettings as any).serverName = oldName;
-    }
   }
 
   private cancelServerCreation(serverToCancel: server.Server): void {
